ciao-controller: Import openstack image API as a ciao API
Copy the openstack image API verbatim in as a ciao API. Subsequent
changes will remove duplicated code and consolidate data structures.
