+++ ciao-cli/main.go
@@ -167,12 +167,6 @@ func buildBlockURL(format string, args ...interface{}) string {
 	return fmt.Sprintf(prefix+format, args...)
 }
 
-func buildImageURL(format string, args ...interface{}) string {
-	prefix := fmt.Sprintf("https://%s:%d/v2/", *controllerURL, *ciaoPort)
-	prefix = fmt.Sprintf("%s%s/", prefix, *tenantID)
-	return fmt.Sprintf(prefix+format, args...)
-}
-
 func sendHTTPRequestToken(method string, url string, values []queryValue, token string, body io.Reader, content string) (*http.Response, error) {
 	req, err := http.NewRequest(method, os.ExpandEnv(url), body)
 	if err != nil {
