+++ ciao-controller/openstack_image.go
@@ -21,26 +21,18 @@ import (
 	"strings"
 	"time"
 
-	"github.com/ciao-project/ciao/ciao-controller/internal/quotas"
+	"github.com/ciao-project/ciao/ciao-controller/api"
 	imageDatastore "github.com/ciao-project/ciao/ciao-image/datastore"
 	"github.com/ciao-project/ciao/ciao-storage"
 	"github.com/ciao-project/ciao/database"
-	"github.com/ciao-project/ciao/openstack/image"
 	"github.com/ciao-project/ciao/payloads"
 	"github.com/ciao-project/ciao/ssntp/uuid"
 	"github.com/golang/glog"
-	"github.com/gorilla/mux"
 	"github.com/pkg/errors"
 )
 
-// ImageService is the context for the image service implementation.
-type ImageService struct {
-	ds imageDatastore.DataStore
-	qs *quotas.Quotas
-}
-
 // CreateImage will create an empty image in the image datastore.
-func (is *ImageService) CreateImage(tenantID string, req image.CreateImageRequest) (image.DefaultResponse, error) {
+func (c *controller) CreateImage(tenantID string, req api.CreateImageRequest) (api.DefaultResponse, error) {
 	// create an ImageInfo struct and store it in our image
 	// datastore.
 	glog.Infof("Creating Image: %v", req.ID)
@@ -51,13 +43,13 @@ func (is *ImageService) CreateImage(tenantID string, req image.CreateImageReques
 	} else {
 		if _, err := uuid.Parse(id); err != nil {
 			glog.Errorf("Error on parsing UUID: %v", err)
-			return image.DefaultResponse{}, image.ErrBadUUID
+			return api.DefaultResponse{}, api.ErrBadUUID
 		}
 
-		img, _ := is.ds.GetImage(tenantID, id)
+		img, _ := c.ids.GetImage(tenantID, id)
 		if img != (imageDatastore.Image{}) {
 			glog.Errorf("Image [%v] already exists", id)
-			return image.DefaultResponse{}, image.ErrAlreadyExists
+			return api.DefaultResponse{}, api.ErrAlreadyExists
 		}
 	}
 
@@ -71,17 +63,17 @@ func (is *ImageService) CreateImage(tenantID string, req image.CreateImageReques
 		Visibility: req.Visibility,
 	}
 
-	err := is.ds.CreateImage(i)
+	err := c.ids.CreateImage(i)
 	if err != nil {
 		glog.Errorf("Error on creating image: %v", err)
-		return image.DefaultResponse{}, err
+		return api.DefaultResponse{}, err
 	}
 
-	res := <-is.qs.Consume(tenantID, payloads.RequestedResource{Type: payloads.Image, Value: 1})
+	res := <-c.qs.Consume(tenantID, payloads.RequestedResource{Type: payloads.Image, Value: 1})
 	if !res.Allowed() {
-		is.ds.DeleteImage(tenantID, id)
-		is.qs.Release(tenantID, payloads.RequestedResource{Type: payloads.Image, Value: 1})
-		return image.DefaultResponse{}, image.ErrQuota
+		c.ids.DeleteImage(tenantID, id)
+		c.qs.Release(tenantID, payloads.RequestedResource{Type: payloads.Image, Value: 1})
+		return api.DefaultResponse{}, api.ErrQuota
 	}
 
 	glog.Infof("Image %v created", id)
@@ -90,12 +82,12 @@ func (is *ImageService) CreateImage(tenantID string, req image.CreateImageReques
 	if len(i.Tags) > 0 {
 		tags = strings.Split(i.Tags, ",")
 	}
-	return image.DefaultResponse{
-		Status:     image.Queued,
+	return api.DefaultResponse{
+		Status:     api.Queued,
 		CreatedAt:  i.CreateTime,
 		Tags:       tags,
 		Locations:  make([]string, 0),
-		DiskFormat: image.Raw,
+		DiskFormat: api.Raw,
 		Visibility: i.Visibility,
 		Self:       fmt.Sprintf("/v2/images/%s", i.ID),
 		Protected:  false,
@@ -107,18 +99,18 @@ func (is *ImageService) CreateImage(tenantID string, req image.CreateImageReques
 	}, nil
 }
 
-func createImageResponse(img imageDatastore.Image) (image.DefaultResponse, error) {
+func createImageResponse(img imageDatastore.Image) (api.DefaultResponse, error) {
 	size := int(img.Size)
 	tags := []string{}
 	if len(img.Tags) > 0 {
 		tags = strings.Split(img.Tags, ",")
 	}
-	return image.DefaultResponse{
+	return api.DefaultResponse{
 		Status:     img.State.Status(),
 		CreatedAt:  img.CreateTime,
 		Tags:       tags,
 		Locations:  make([]string, 0),
-		DiskFormat: image.DiskFormat(img.Type),
+		DiskFormat: api.DiskFormat(img.Type),
 		Visibility: img.Visibility,
 		Self:       fmt.Sprintf("/v2/images/%s", img.ID),
 		Protected:  false,
@@ -131,11 +123,11 @@ func createImageResponse(img imageDatastore.Image) (image.DefaultResponse, error
 }
 
 // ListImages will return a list of all the images in the datastore.
-func (is *ImageService) ListImages(tenant string) ([]image.DefaultResponse, error) {
+func (c *controller) ListImages(tenant string) ([]api.DefaultResponse, error) {
 	glog.Infof("Listing images from [%v]", tenant)
-	response := []image.DefaultResponse{}
+	response := []api.DefaultResponse{}
 
-	images, err := is.ds.GetAllImages(tenant)
+	images, err := c.ids.GetAllImages(tenant)
 	if err != nil {
 		glog.Errorf("Error on retrieving images from tenant [%v]: %v", tenant, err)
 		return response, err
@@ -150,11 +142,11 @@ func (is *ImageService) ListImages(tenant string) ([]image.DefaultResponse, erro
 }
 
 // UploadImage will upload a raw image data and update its status.
-func (is *ImageService) UploadImage(tenantID, imageID string, body io.Reader) (image.NoContentImageResponse, error) {
+func (c *controller) UploadImage(tenantID, imageID string, body io.Reader) (api.NoContentImageResponse, error) {
 	glog.Infof("Uploading image: %v", imageID)
-	var response image.NoContentImageResponse
+	var response api.NoContentImageResponse
 
-	err := is.ds.UploadImage(tenantID, imageID, body)
+	err := c.ids.UploadImage(tenantID, imageID, body)
 	if err != nil {
 		glog.Errorf("Error on uploading image: %v", err)
 		return response, err
@@ -166,17 +158,17 @@ func (is *ImageService) UploadImage(tenantID, imageID string, body io.Reader) (i
 }
 
 // DeleteImage will delete a raw image and its metadata
-func (is *ImageService) DeleteImage(tenantID, imageID string) (image.NoContentImageResponse, error) {
+func (c *controller) DeleteImage(tenantID, imageID string) (api.NoContentImageResponse, error) {
 	glog.Infof("Deleting image: %v", imageID)
-	var response image.NoContentImageResponse
+	var response api.NoContentImageResponse
 
-	err := is.ds.DeleteImage(tenantID, imageID)
+	err := c.ids.DeleteImage(tenantID, imageID)
 	if err != nil {
 		glog.Errorf("Error on deleting image: %v", err)
 		return response, err
 	}
 
-	is.qs.Release(tenantID, payloads.RequestedResource{Type: payloads.Image, Value: 1})
+	c.qs.Release(tenantID, payloads.RequestedResource{Type: payloads.Image, Value: 1})
 
 	response.ImageID = imageID
 	glog.Infof("Image %v deleted", imageID)
@@ -184,11 +176,11 @@ func (is *ImageService) DeleteImage(tenantID, imageID string) (image.NoContentIm
 }
 
 // GetImage will get the raw image data
-func (is *ImageService) GetImage(tenantID, imageID string) (image.DefaultResponse, error) {
+func (c *controller) GetImage(tenantID, imageID string) (api.DefaultResponse, error) {
 	glog.Infof("Getting Image [%v] from [%v]", imageID, tenantID)
-	var response image.DefaultResponse
+	var response api.DefaultResponse
 
-	img, err := is.ds.GetImage(tenantID, imageID)
+	img, err := c.ids.GetImage(tenantID, imageID)
 	if err != nil {
 		glog.Errorf("Error on getting image: %v", err)
 		return response, err
@@ -196,7 +188,7 @@ func (is *ImageService) GetImage(tenantID, imageID string) (image.DefaultRespons
 
 	if (img == imageDatastore.Image{}) {
 		glog.Infof("Image %v not found", imageID)
-		return response, image.ErrNoImage
+		return response, api.ErrNoImage
 	}
 
 	response, _ = createImageResponse(img)
@@ -205,7 +197,7 @@ func (is *ImageService) GetImage(tenantID, imageID string) (image.DefaultRespons
 }
 
 // Init initialises the image service
-func (is *ImageService) Init(qs *quotas.Quotas) error {
+func (c *controller) InitImageDatastore() error {
 	dbDir := filepath.Dir(*imageDatastoreLocation)
 	dbFile := filepath.Base(*imageDatastoreLocation)
 
@@ -258,9 +250,8 @@ func (is *ImageService) Init(qs *quotas.Quotas) error {
 	glog.Infof("RawDataStore  : %T", config.RawDataStore)
 	glog.Infof("MetaDataStore : %T", config.MetaDataStore)
 
-	is.ds = &imageDatastore.ImageStore{}
-	is.qs = qs
-	err = is.ds.Init(config.RawDataStore, config.MetaDataStore)
+	c.ids = &imageDatastore.ImageStore{}
+	err = c.ids.Init(config.RawDataStore, config.MetaDataStore)
 	if err != nil {
 		return err
 	}
@@ -281,15 +272,4 @@ type ImageConfig struct {
 
 	// MetaDataStore is an interface to a persistent datastore for the image meta data.
 	MetaDataStore imageDatastore.MetaDataStore
-}
-
-func (c *controller) createImageRoutes(r *mux.Router) error {
-	apiConfig := image.APIConfig{
-		ImageService: c.is,
-	}
-
-	// get our routes.
-	image.Routes(apiConfig, r)
-
-	return nil
 }
